4 undefined Bands fix
git-svn-id: https://svn.java.net/svn/imageio-ext~svn/branches/1.0.x@1097 866bbf7d-4862-ccf5-812a-c90ffd63a74d

+++ plugin/kakadujp2/src/main/java/it/geosolutions/imageio/plugins/jp2k/box/ChannelDefinitionBoxMetadataNode.java
@@ -45,7 +45,7 @@ public class ChannelDefinitionBoxMetadataNode extends BaseJP2KBoxMetadataNode {
         numChannels = box.getNum();
         final short[] channel = box.getChannel();
         final short[] assoc = box.getAssociation();
-        final short[] type = box.getTypes();
+        final int[] type = box.getTypes();
         IIOMetadataNode child = new IIOMetadataNode("NumberOfChannelDefinition");
         child.setUserObject(new Short(numChannels));
         numberOfChannelDefinitions = Short.toString(numChannels);
@@ -66,8 +66,8 @@ public class ChannelDefinitionBoxMetadataNode extends BaseJP2KBoxMetadataNode {
             child.appendChild(child1);
 
             child1 = new IIOMetadataNode("ChannelType");
-            child1.setUserObject(new Short(type[i]));
-            types[i] = Short.toString(type[i]);
+            child1.setUserObject(new Integer(type[i]));
+            types[i] = Integer.toString(type[i]);
             child1.setNodeValue(types[i]);
             child.appendChild(child1);
